refactor: rename all component files to kebab-case following shadcn convention
- Rename BaseGrid.tsx -> base-grid.tsx
- Rename BaseTopBar.tsx -> base-top-bar.tsx
- Rename DashboardTopBar.tsx -> dashboard-top-bar.tsx
- Rename CreateGalleryModal.tsx -> create-gallery-modal.tsx
- Rename FolderSelector.tsx -> folder-selector.tsx
- Rename ForeverStorageProgressModal.tsx -> forever-storage-progress-modal.tsx
- Rename GalleryCard.tsx -> gallery-card.tsx
- Rename GalleryGrid.tsx -> gallery-grid.tsx
- Rename GalleryList.tsx -> gallery-list.tsx
- Rename GalleryTopBar.tsx -> gallery-top-bar.tsx
- Rename ItemUploadButton.tsx -> item-upload-button.tsx
- Rename MemoryActions.tsx -> memory-actions.tsx
- Rename MemoryCard.tsx -> memory-card.tsx
- Rename MemoryGrid.tsx -> memory-grid.tsx
- Rename MemoryStatus.tsx -> memory-status.tsx
- Rename ShareDialog.tsx -> share-dialog.tsx

Update all import paths to use new kebab-case filenames
Maintain PascalCase component names (React convention)
Follow shadcn/ui file naming best practices
