+++ src/components/galleries/folder-selector.tsx
