+++ src/components/galleries/gallery-card.tsx
